Initial Commit, refactored classes from the Program.cs. Renamed Program.cs to Main.cs. Created class for start screen menu.
