+++ Main.cs
@@ -1,0 +1,20 @@
+﻿using System;
+
+namespace SpaceGame
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Space Game!");
+            Character playerCharacter = new Character("Luis", 3);
+            Console.WriteLine($"Name:{playerCharacter.GetName()}");
+            Console.WriteLine($"Current Currency:{playerCharacter.GetCurrency()}");
+            playerCharacter.AddCurrency(15);
+            Console.WriteLine($"Current Currency:{playerCharacter.GetCurrency()}");
+            playerCharacter.SubtractCurrency(7);
+            Console.WriteLine($"Current Currency:{playerCharacter.GetCurrency()}");
+        }
+        
+    }
+}
