+++ StartScreenMenu.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using SpaceGame;
+
+/// <summary>
+/// Summary description for Class1
+/// </summary>
+public class StartScreenMenu
+{
+	public void Choices()
+	{
+        Console.WriteLine("Welcome to the Space Game!");
+        Console.WriteLine("Make a choice!");
+		
+
+	}
+}
